Add ManagementDataTransformerTest
RISDEV-7401
